requirements.txt
pandas>=2.0.0
Pillow>=11.0.0 